rpc: extract shared wire test codec
